requirements.txt
flask
requests
json2html
cassandra-driver
flask-cassandra
